pyproject.toml
[project]
name = "realtime-indexer-qa-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.3.0",
    "llama-index==0.9.38",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "pathway==0.8.4",
    "python-dotenv==1.0.1",
    "streamlit==1.31.0",
    "torch>=2.9.1",
    "traceloop-sdk==0.12.5",
    "transformers>=4.57.3",
    "pypdf>=4.0.0",
    "python-docx>=0.8.0",
    "openpyxl>=3.0.0",
]
